Refactor Extension to be Plugin, extensions -> extend
